fixing warnings, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@16608 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSVehicle.cpp
@@ -2072,7 +2072,7 @@ MSVehicle::getBestLaneOffset() const {
 void
 MSVehicle::adaptBestLanesOccupation(int laneIndex, SUMOReal density) {
     std::vector<MSVehicle::LaneQ>& preb = myBestLanes.front();
-    assert(laneIndex < preb.size());
+    assert(laneIndex < (int)preb.size());
     preb[laneIndex].occupation = density + preb[laneIndex].nextOccupation;
 }
